requirements.txt
ptyprocess>=0.5.1
tornado>=4.1
docker-py>=1.7.1
pytest>=3.0.2